% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{block_alignments}
\alias{block_alignments}
\title{Makes a ggplot2 figure of alignment blocks arranged by reads along target}
\usage{
block_alignments(df)
}
\arguments{
\item{df}{A data_frame output from make_block_df}
}
\value{
A ggplot2 object
}
\description{
Makes a ggplot2 figure of alignment blocks arranged by reads along target
}

