% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{make_block_df}
\alias{make_block_df}
\title{Takes data_frame output from read_blitz_tsv and returns a GRanges object}
\usage{
make_block_df(df)
}
\arguments{
\item{df}{data_frame output from read_blitz_tsv}
}
\value{
A GRanges objects
}
\description{
Takes data_frame output from read_blitz_tsv and returns a GRanges object
}

