% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{make_block_rows}
\alias{make_block_rows}
\title{Takes a row form a data_frame output from read_blitz_tsv and expands list columns into rows}
\usage{
make_block_rows(row)
}
\arguments{
\item{row}{A row from a data_frame, output from read_blitz_tsv}
}
\value{
A data_frame
}
\description{
Takes a row form a data_frame output from read_blitz_tsv and expands list columns into rows
}

