% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{make_gr_rows}
\alias{make_gr_rows}
\title{Takes a row from a data_frame output from read_blitz_tsv, converts into a GenomicRanges object}
\usage{
make_gr_rows(row)
}
\arguments{
\item{row}{A row from a data_frame, output from read_blitz_tsv}
}
\value{
A GenomicRanges object
}
\description{
Takes a row from a data_frame output from read_blitz_tsv, converts into a GenomicRanges object
}

