% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_num_list}
\alias{make_num_list}
\title{Make a numeric vector from a string of comma separated numbers}
\usage{
make_num_list(num_string)
}
\arguments{
\item{num_string}{A string vector of comma separated numbers without spaces}
}
\description{
Make a numeric vector from a string of comma separated numbers
}

