% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{read_blitz_tsv}
\alias{read_blitz_tsv}
\title{Read a blitz output tsv file}
\usage{
read_blitz_tsv(path)
}
\arguments{
\item{path}{The path to a blitz tsv}
}
\value{
A data_frame, where blockSizes, qStarts and tStarts are stored as lists
}
\description{
Read a blitz output tsv file
}

